Add files via upload
This project presents an overview of phishing attacks, covering their types:like email phishing,how to recognize them, and the risks of falling victim. The slides are supported with interactive quizzes that use real-life scenarios and multiple-choice questions to test understanding on spotting phishing emails, avoiding and responding if targeted.